requirements.txt
python-dateutil==2.4.2
PyVimeo==0.3.0
